oWasp - We wree creating some SQL code on the IIS server when adding components and expressions. This was open to some SQL injection, so I've swapped it all out for some paramterised stored procs.
As part of this I've also put the expression code files into an expression namespace.

+++ Components/COAInt_Server/Expressions/modExpression.vb
@@ -201,7 +201,7 @@ ErrorTrap:
 
 	End Function
 
-	Public Function ValidateFunctionParameters(ByRef plngFunctionID As Integer, ByRef piResultType As ExpressionValueTypes, Optional ByRef piParam1Type As Short = 0, Optional ByRef piParam2Type As Short = 0, Optional ByRef piParam3Type As Short = 0, Optional ByRef piParam4Type As Short = 0, Optional ByRef piParam5Type As Short = 0, Optional ByRef piParam6Type As Short = 0) As Boolean
+	Public Function ValidateFunctionParameters(ByRef plngFunctionID As Integer, ByRef piResultType As ExpressionValueTypes, Optional ByRef piParam1Type As Integer = 0, Optional ByRef piParam2Type As Integer = 0, Optional ByRef piParam3Type As Integer = 0, Optional ByRef piParam4Type As Integer = 0, Optional ByRef piParam5Type As Integer = 0, Optional ByRef piParam6Type As Integer = 0) As Boolean
 		' Validate the given function with the given parameters.
 		' Return the result type in the piResultType parameter.
 		On Error GoTo ErrorTrap
